commented login functions
