+++ app/Http/Controllers/Article/ArticleController.php
@@ -15,36 +15,53 @@ use App\Events\NewArticleAdded;
 use App\Events\NewCommentAdded;
 use App\Helper\Slug;
 
+/**
+ * handles all article related actions
+ */
 class ArticleController  extends Controller
 {
+    //to upload image to the server
     use ImageUpload;
 
+    /**
+     * view fullscreen article 
+     * @param slug name of the article to viewed
+     * @return view of full page article
+     */
     public function index($id)
     {
+        //finds article by its slug name
         $article=ArticleModel::where('slug',$id)->first();
         if($article){
+            //increment number of views of thr article
             $article->timestamps=false;
             $article->increment('views');
             if($article->is_published){
+                //normal user can view article if it is published
                 return view('post.full',['article'=>$article]);
             }else if(Gate::allows('update-post', $article)){
+                //athorized user can view article even if it is not published
                 return view('post.full',['article'=>$article,'status'=>'unpublished']);
             }else{
+                //return to error page with error mesaage
                 return view('error',['message'=>'article is not published yet']);
             }
         }
         return view('error',['message'=>'article does not exist']);
     }
     /**
-     * get add article form
+     * get add article form for adding new article
      */
     public function getAddForm()
     {
+        //gets all catagory list
         $catagory=Catagory::all();
         return view('post.add',['catagory'=>$catagory]);
     }
     /**
-     * Edit an article
+     * Gets edit article form for editing the article
+     * @param slug for the article to be edited
+     * @return view edit page
      */
     public function editForm($id)
     {
@@ -60,6 +77,11 @@ class ArticleController  extends Controller
         }
     }
 
+    /**
+     * update article when user submits edit form
+     * @param request object
+     * @return redirect with success message
+     */
     public function edit(Request $request)
     {
         $this->validator($request);
@@ -69,6 +91,7 @@ class ArticleController  extends Controller
         }
         $catagory=Catagory::all();
         if(Gate::allows('update-post', $article)){
+            //only aithorized user can edit the post
             if( $request->image){
                 try {
                 $filePath = $this->UserImageUpload($request->image); //Passing $request->image as parameter to our created method
@@ -78,6 +101,7 @@ class ArticleController  extends Controller
                 return redirect()->back()->with('status','failed');
                 }
             }
+            //update article details
             $article->title = $request->title;
             $article->slug = Slug::createSlug('article',$request->title);
             $article->body = $request->body;
@@ -90,18 +114,27 @@ class ArticleController  extends Controller
                 }
             }  
             $article->save();
+            //redirect to te article page when update is successfull
             return redirect(route('article',$article->slug))->with('status','success');
-
         }else{
+            //redirect to error page with custom message
             return view('error',['message'=>'you are not autherorized']);
         }
-
     }
 
+    /**
+     * delete article when user deletes the article
+     * deleting an article will also delete all likes and comments
+     * @param slug name of the article to be deleted
+     * @return redirect with confirmation
+     */
     public function delete($id)
     {
+
         $article= ArticleModel::where('slug',$id)->first();
         if(Gate::allows('update-post', $article)){
+            //only authorized user can dlete the article
+            //remove all likes,comment and catagory before deleting the article
             $article->likes()->detach();
             $article->comments()->detach();
             $article->catagories()->detach();
@@ -113,60 +146,40 @@ class ArticleController  extends Controller
     }
 /**
  * like or dislike an article by its id
+ * any registered user can like an article
+ * a single user can only like once an article
+ * guest user can't like any article
+ * @param id of the article to be liked
+ * @return response with like status
  */
     public function like($id)
     {
-        # code...
         if(!Auth::check() || Auth::id() === 0){
+            //guest user can't like a post
             return ['error'=>true,'message'=>'please login first'];
         }
         $article=ArticleModel::find($id);
         if(is_null($article)){
+            //invalid article if id doesn't exist
             return ['error'=>true,'message'=>'invalid article id'];  
         }
         $iliked=$article->likes()->where('user_id',Auth::id())->get();
+        //check user liked a post or not
         if(count($iliked)>0){
             $article->likes()->detach(Auth::id());
             return ['error'=>false,'liked'=>false];
         }
         try{
-
+            //like post if user didn't liked already
             $article->likes()->attach(Auth::id(),['react'=>'LIKE']);
             return ['error'=>false,'liked'=>true];
         }catch(Exception $e){
+            //invalid response if already liked
             return ['error'=>true,'message'=>'already liked'];
         }
         
     }
 
-    public function comment(Request $request)
-    {
-        # code...
-        $request->validate([
-            'id'=>['required','string'],
-            'comment'=>['required','string']
-        ]);
-        $article= ArticleModel::find($request->id);
-        if(is_null($article)){
-            return view('error')->with(['message'=>'invalid article']);
-        }   
-        if(Auth::check()){
-            if(Auth::id()==0){
-                $comment = new Comments(['article_id'=>$article->id,'is_published'=>0,'body'=>$request->comment,'user_id'=>Auth::id()]);
-                $comment->save();
-                event(new NewCommentAdded($comment));
-                
-                return redirect()->back()->with(['comment'=>'pending for admin approval']);
-            }else{
-                $comment = new Comments(['article_id'=>$article->id,'body'=>$request->comment,'user_id'=>Auth::id()]);
-                $comment->save();
-                event(new NewCommentAdded($comment));
-            }
-            return redirect()->back()->with(['status'=>'success']);
-        }else{
-            return view('error')->with(['message'=>'article']);
-        }
-    }
     /**
      * validate inpu request 
      */
